cmakeGTK3: move from cross compilation to native with QEMU
Cross compilation had some issues with the GTK 4 template due to Debian
dependencies. To keep the GTK 3 and GTK 4 templates as close as
possible, move both to using native compilation (with the help of QEMU
emulation).

This also solves an issue with 32-bit Arm cross-compilation specific to
the IDE Templates variable substitution.

Signed-off-by: Leonardo Graboski Veiga <leonardo.veiga@toradex.com>